requirements.txt
playwright>=1.40.0
openai>=1.3.5
APScheduler>=3.10.4
python-dotenv>=1.0.0
